refactors api controller to inherit from ActionController::API creates new helper for ApiController removes redundant code
