+++ app/controllers/dashboard_controller.rb
@@ -5,7 +5,6 @@ class DashboardController < ApplicationController
     @resources = Resource.all
   end
 
-
   def create_booking
     id = params[:booking][:resource_id]
     @resource = Resource.find(id)
